requirements.txt
PyYAML>=6.0
boto3>=1.21.27